@@ -5,6 +5,7 @@
   "main": "index.js",
   "scripts": {
     "start": "webpack-dev-server --open --mode development",
+    "dev": "npm run start -dev",
     "build": "webpack --mode production",
     "test": "jest",
     "test:coverage": "jest --coverage",
@@ -15,7 +16,6 @@
     "setupFilesAfterEnv": [
       "<rootDir>/src/setupTest.js"
     ],
-
     "collectCoverage": true,
     "moduleNameMapper": {
       "\\.(jpg|ico|jpeg|png|gif|eot|otf|webp|svg|ttf|woff|woff2|mp4|webm|wav|mp3|m4a|aac|oga)$": "<rootDir>/__mocks__/fileMock.js",
@@ -41,7 +41,6 @@
     "@babel/preset-env": "^7.5.0",
     "@babel/preset-react": "^7.0.0",
     "babel-loader": "^8.0.6",
-    "check-prop-types": "^1.1.2",
     "clean-webpack-plugin": "^3.0.0",
     "coveralls": "^3.0.5",
     "css-loader": "^3.0.0",
@@ -49,8 +48,9 @@
     "enzyme-adapter-react-16": "^1.14.0",
     "eslint": "^5.3.0",
     "eslint-config-airbnb": "^17.1.1",
+    "eslint-import-resolver-webpack": "^0.11.1",
     "eslint-loader": "^2.2.1",
-    "eslint-plugin-import": "^2.18.0",
+    "eslint-plugin-import": "^2.18.2",
     "eslint-plugin-jsx-a11y": "^6.2.3",
     "eslint-plugin-react": "^7.14.2",
     "file-loader": "^4.0.0",
@@ -59,19 +59,26 @@
     "identity-obj-proxy": "^3.0.0",
     "jest": "^24.8.0",
     "jest-enzyme": "^7.0.2",
-    "prop-types": "^15.7.2",
     "style-loader": "^0.23.1",
     "webpack": "^4.35.2",
     "webpack-cli": "^3.3.5",
     "webpack-dev-server": "^3.7.2"
   },
   "dependencies": {
+    "@babel/polyfill": "^7.4.4",
     "@fortawesome/fontawesome-svg-core": "^1.2.19",
     "@fortawesome/free-solid-svg-icons": "^5.9.0",
     "@fortawesome/react-fontawesome": "^0.1.4",
-    "react": "^16.8.6",
+    "axios": "^0.19.0",
+    "babel-eslint": "^10.0.2",
+    "check-prop-types": "^1.1.2",
+    "jwt-decode": "^2.2.0",
+    "prop-types": "^15.7.2",
+    "react": "^16.9.0",
     "react-dom": "^16.8.6",
     "react-redux": "^7.1.0",
+    "react-router-dom": "^5.0.1",
+    "react-toastify": "^5.3.2",
     "redux": "^4.0.3",
     "redux-thunk": "^2.3.0"
   }
