@@ -1,24 +1,25 @@
-import React from 'react'
-import PropTypes from 'prop-types'
+import React from 'react';
+import PropTypes from 'prop-types';
 
-const Button = ({name, value, onClick, className}) => {
-  return (
-    <button 
-      name={name} 
-      onClick={onClick}
-      className={className}
-      data-test="btn"
-      >
-      {value}
-    </button>
-  )
-}
+const Button = ({
+  name, value, onClick, className,
+}) => (
+  <button
+    name={name}
+    onClick={onClick}
+    className={className}
+    data-test="btn"
+    type="button"
+  >
+    {value}
+  </button>
+);
 
 Button.propTypes = {
   name: PropTypes.string.isRequired,
-  value: PropTypes.string.isRequired,
-  onClick: PropTypes.func,
-  className: PropTypes.string.isRequired
-}
+  onClick: PropTypes.func.isRequired,
+  value: PropTypes.any.isRequired,
+  className: PropTypes.string.isRequired,
+};
 
 export default Button;
