@@ -1,13 +1,8 @@
-import React from "react";
-import ReactDOM from "react-dom";
-import { Provider } from 'react-redux';
-import store from './reducers/store';
+import React from 'react';
+import ReactDOM from 'react-dom';
+import '@babel/polyfill';
+import App from './App';
 
-import App from './App'
-import "./assets/app.css";
+import './assets/app.css';
 
-ReactDOM.render(
-    <Provider store={store}>
-      <App />
-    </Provider>,
-    document.getElementById("root"));
+ReactDOM.render(<App />, document.getElementById('root'));
