@@ -1,7 +1,14 @@
 import { library } from '@fortawesome/fontawesome-svg-core';
-import { faAt, faLock } from '@fortawesome/free-solid-svg-icons';
+import {
+  faAt,
+  faLock,
+  faUser,
+  faSpinner,
+} from '@fortawesome/free-solid-svg-icons';
 
 library.add(
   faAt,
-  faLock
+  faLock,
+  faUser,
+  faSpinner,
 );
